requirements.txt
asgiref==3.7.2
black==24.1.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django==4.2.9
django-environ==0.11.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.45.1
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
zipp==3.17.0
